B-02386: Support for removing content from section in publish content wizard.

+++ modules/cms-core/src/main/java/com/enonic/cms/core/structure/menuitem/MenuItemServiceImpl.java
@@ -413,7 +413,7 @@ public class MenuItemServiceImpl
         final ContentHomeEntity contentHome =
             contentHomeDao.findByKey( new ContentHomeKey( section.getSite().getKey(), content.getKey() ) );
 
-        if ( contentHome != null )
+        if ( contentHome != null && section.getKey() == contentHome.getMenuItem().getKey())
         {
             content.removeContentHome( section.getSite().getKey() );
             contentHomeDao.delete( contentHome );
